Add circular dependencies check
