+++ .github/workflows/check.yml
@@ -34,3 +34,10 @@ jobs:
       - run: npm i
       - run: npm run build
       - run: npm run test:types
+
+  circular:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - run: npm i
+      - run: npm run circular
